+++ RELEASES.md
@@ -522,3 +522,4 @@
 ----------------
   - fix: issue with __transformResponse protocol happn_1, bad callback
   - fix: security, users that are substrings of other users get returned by getUser
+  - test: updated travis.yml